msm: apuclock-krait: Introduce a simple VDD interface
Use the newly added CPUFreq API to create a voltage-on-frequency table
and make it accessible to users.  This can help with a stress balancing
to improve CPU stability, especially on floor frequencies, which can be
used quite often.

Krait power subsystem uses vdd_core variable to set a CPU core voltage.
Now this value can be changed manually from userspace. acpuclk-krait
returns the VDD table in the style of "frequency: voltage".

 ...
 1190mhz: 885 mV
 1267mhz: 895 mV
 1497mhz: 930 mV
 ...

Voltage values are writable and can be changed by passing the whole VDD
table to UV_mV_table sysfs node.  That's not convenient for a terminal
editing, but is evaluated by third-party applications which expect exactly
this implementation.

Both sysfs functions report only the actual frequencies used, leaving the
disabled ones alone.  In addition, write function ensures that the passed
values are within the bounds.  Lower VDD limit comes straight from Krait
Regulator when the upper one is equal to maximum voltage in acpuclk data.

Change-Id: I79c279523268953b72139e9480e00b31ddbf1a36
Signed-off-by: Alex Saiko <solcmdr@gmail.com>